[ALPS03226208] secmem: Fix build error for 64bit format string
[Detail]
phys_addr_t depends on CONFIG_PHYS_ADDR_T_64BIT and can be u32 or u64.
Cast to u64 for both 32/64 configuration.

MTK-Commit-Id: 83ee561a7c5da11139f5d55de9dc77adcbb12b0c

CR-Id: ALPS03226208
Feature: Google Widevine Level1
(cherry picked from commit c657dbf392a8e9577de2c76722e4841762f5ddff)

+++ drivers/misc/mediatek/secmem/secmem.c
@@ -517,7 +517,7 @@ static int secmem_region_alloc(void)
 	}
 
 	if (pa == 0 || size == 0) {
-		MSG(ERR, "%s: invalid pa(0x%llx) or size(0x%lx)\n", __func__, pa, size);
+		MSG(ERR, "%s: invalid pa(0x%llx) or size(0x%lx)\n", __func__, (u64)pa, size);
 		return -1;
 	}
 
@@ -538,7 +538,7 @@ static int secmem_region_alloc(void)
 #endif
 
 	INFO("phyaddr=0x%llx sz=0x%lx region_online=%u region_ref=%u\n",
-			pa, size, secmem_region_online, secmem_region_ref);
+			(u64)pa, size, secmem_region_online, secmem_region_ref);
 
 	return 0;
 }
